fix(review): resolve 7 test failures from initial review fix commit
- Fix ARCH-003: add logger.debug to suppress-close except in CampaignStateMutator.__exit__
- Fix never-raises contract: restore top-level try/except Exception in
  _write_dispatch_to_campaign_state (arch test requires it as first statement)
- Fix layer boundary: replace autoskillit.cli.fleet import in TestFlockLockTarget
  with filesystem path navigation (fleet_root.parent / "cli" / "fleet")
- Fix layer boundary: extract TestReapAndResumeMutualExclusion to
  tests/integration/test_fleet_concurrency.py (CLI import forbidden in tests/fleet/)
- Fix monkeypatch target: revert is_dispatch_session_alive to import via
  fleet.__init__ so existing monkeypatches on autoskillit.fleet still work
- Fix TestClassifyStaleDispatch: use empty sidecar files so raw_lines=[] triggers
  the RESUMABLE/INTERRUPTED decision path correctly; remove broken patch.object
